requirements.txt
fastapi
uvicorn
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langgraph>=0.0.10
huggingface_hub
transformers
psycopg2-binary
requests
redis
pydantic>=2.0.0
aiohttp
weaviate-client>=4.0.0
tenacity>=8.0.0
tqdm
python-multipart
jinja2
python-dotenv